requirements.txt
streamlit==1.30.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.0
streamlit-faker==0.0.3
streamlit-folium==0.18.0
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.3
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
Pillow ==10.0.1
matplotlib ==3.8.0
matplotlib-inline==0.1.6
numpy==1.26.3
plotly==5.18.0
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
python-dateutil==2.8.2
pprint
pandas==2.1.4

